tunescout_api/requirements.txt
blinker==1.9.0
click==8.3.1
clickhouse-driver==0.2.10
contourpy==1.3.3
cycler==0.12.1
ffmpeg-python==0.2.0
Flask==3.1.2
fonttools==4.61.0
future==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mysql-connector-python==9.5.0
numpy==2.3.5
packaging==25.0
pillow==12.0.0
pydub==0.25.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.16.3
six==1.17.0
tzlocal==5.3.1
Werkzeug==3.1.4
